apps/02-chat-with-context/requirements.txt
streamlit==1.44.0
langchain-core==0.3.59
google-generativeai==0.8.4
python-dotenv==1.1.0